[jsc] Migrate, where possible, to using TZoneMalloc in Source/JavaScriptCore https://bugs.webkit.org/show_bug.cgi?id=296828 rdar://157329141
Reviewed by NOBODY (OOPS!).

As part of our effort to deprecate MAKE_FAST_ALLOCATED, any objects that
can be allocated using TZoneMalloc should be made to do so. Not all
types can be so moved: e.g. variable-sized types are generally excluded.
This patch moves all types which can use TZones over to using TZones.
Qualifying types in the rest of the repository will be moved in a later
patch.

One notable omission is handling for the WebInspector codegen classes,
which do create a few FastMalloc-allocated types. I made a first pass at
converting those as well but ran into strange issues which persuaded me
to leave that for later so as to get this change, which is
rebase-sensitive, out of the door.

The other big category I had to skip over was that of types tagged with
WTF_DEPRECATED_MAKE_FAST_ALLOCATED_WITH_HEAP_IDENTIFIER, as the
equivalent functionality is not yet implemented in TZoneMalloc (and
doing so would be out of scope for this patch).