[First-Letter] Fix ::first-letter disappearing after style change. https://bugs.webkit.org/show_bug.cgi?id=297570 <rdar://145550507>
Reviewed by NOBODY (OOPS!).

Changing a block with a ::first-letter renderer incorrectly destroys the ::first-letter renderer.
This would remove the first letter from the rendered output each time we update the text due
to style changes.

This PR fixes that bug by by removing the code that destroys the ::first-letter renderer in
`RenderTextFragment::setTextInternal` as it should not be responsible for destroying it.

* LayoutTests/editing/first-letter-text-transform-bug-expected.txt: Added.
* LayoutTests/editing/first-letter-text-transform-bug.html: Added.
* Source/WebCore/rendering/RenderTextFragment.cpp:
(WebCore::RenderTextFragment::setTextInternal):