Disable WasmMemoryToBufferAPI on 32-bit Need the bug URL (OOPS!).
Reviewed by NOBODY (OOPS!).

This PR prevents JSC from running in 32-bit when using the option
--useWasmMemoryToBufferAPIs=true, and disables the tests for 32-bit if this
feature is required.

useWasmMemoryToBufferAPIs enables resizable memory buffer, but the spec [0]
defines that the max buffer length should be 4GB, but we can't create a 4GB
buffer in 32-bit JSC, the max size for arrays is 4GB - 1 byte.

[0] https://webassembly.github.io/threads/js-api/index.html#create-a-resizable-memory-buffer

* JSTests/wasm/js-api/memory-toFixedLengthBuffer.js:
(assertSharedFixedLengthBufferOfPageSize):
* JSTests/wasm/js-api/memory-toResizableBuffer.js:
* Source/JavaScriptCore/runtime/Options.cpp:
(JSC::Options::assertOptionsAreCoherent):