dNR: make sure every rule ID is unique https://bugs.webkit.org/show_bug.cgi?id=298070 rdar://72159785
Reviewed by NOBODY (OOPS!).

This simple patch reports an error when a static dNR ruleset contains a duplicate identifier. It's
important that we do this so that developers are aware that they might be getting incorrect
information from onRuleMatchedDebug.

The error is surfaced in Safari's extensions preferences pane.

* Source/WebCore/en.lproj/Localizable.strings:
* Source/WebKit/UIProcess/Extensions/Cocoa/WebExtensionContextCocoa.mm:
(WebKit::WebExtensionContext::compileDeclarativeNetRequestRules):
* Source/WebKit/UIProcess/Extensions/Cocoa/_WKWebExtensionDeclarativeNetRequestRule.mm:
* Source/WebKit/UIProcess/Extensions/Cocoa/_WKWebExtensionDeclarativeNetRequestTranslator.mm:
(WebKit::for):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtensionAPIDeclarativeNetRequest.mm:
(TestWebKitAPI::TEST(WKWebExtensionAPIDeclarativeNetRequest, DuplicatedRuleIDs)):