WebKit does not forward debug/info logging from the WebContent process https://bugs.webkit.org/show_bug.cgi?id=298087 rdar://152591955
Reviewed by NOBODY (OOPS!).

We used to only forward debug/info logging from the WebContent processes in
debug builds. While this may be OK for WebKit developers, this is insufficient
for underlying framework developers as they would generally use a release build
of WebKit.

To address the issue, we now rely on `os_log_debug_enabled(OS_LOG_DEFAULT)`
instead to determine if we should forward debug/info messages or not. This is
true when the developer is streaming debug logs or debug logging has been
explicitly enabled system wide (not the default).

* Source/WebKit/Shared/WebProcessCreationParameters.h:
* Source/WebKit/Shared/WebProcessCreationParameters.serialization.in:
* Source/WebKit/UIProcess/Cocoa/WebProcessPoolCocoa.mm:
(WebKit::WebProcessPool::platformInitializeWebProcess):
* Source/WebKit/WebProcess/WebProcess.h:
* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess):
(WebKit::registerLogClient):
(WebKit::WebProcess::initializeLogForwarding):