Layout Test security/contentSecurityPolicy/plugins-types-allows-youtube-plugin-replacement.html is flaky https://bugs.webkit.org/show_bug.cgi?id=173742 rdar://42793011
Reviewed by NOBODY (OOPS!).

* LayoutTests/security/contentSecurityPolicy/plugins-types-allows-youtube-plugin-replacement.html:
The test was flaky because it was relying on a 100ms timer to check if something was loaded.
We instead now rely on a console logging listener since the load in question generates such logging
and it was the presence of this logging that was causing the flakiness.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::addConsoleMessage):
(WebCore::Document::setConsoleMessageListener): Deleted.
* Source/WebCore/dom/Document.h:
* Source/WebCore/page/PageConsoleClient.cpp:
(WebCore::PageConsoleClient::setConsoleMessageListener):
(WebCore::PageConsoleClient::addMessage):
(WebCore::PageConsoleClient::messageWithTypeAndLevel):
* Source/WebCore/page/PageConsoleClient.h:
* Source/WebCore/testing/Internals.cpp:
(WebCore::Internals::resetToConsistentState):
(WebCore::Internals::Internals):
(WebCore::Internals::setConsoleMessageListener):
Fix issue where the console message listener was not reliably called for all console
messages.