Filter primitive subregion with relative values should properly render referenced elements https://bugs.webkit.org/show_bug.cgi?id=298540 rdar://154363710
Reviewed by NOBODY (OOPS!).

SVG filters can be defined independently and referenced via CSS
or another SVG. In this scenario, when a filter primitive uses a
percentage-based subregion, the current implementation fails to render
the referenced element due to empty viewport information.

To solve this, we fallback to object bounding box behavior and use the
given viewport size.

* LayoutTests/svg/filters/svg-filter-external-references-expected.html: Added.
* LayoutTests/svg/filters/svg-filter-external-references.html: Added.
* Source/WebCore/svg/SVGLengthContext.cpp:
(WebCore::dimensionForLengthMode):
(WebCore::SVGLengthContext::resolveRectangle):