[WebXR] Initiaize left & right viewports on Cocoa platforms https://bugs.webkit.org/show_bug.cgi?id=299161 rdar://160918170
Reviewed by NOBODY (OOPS!).

On Cocoa platforms, the viewports are initially 0 x 0. This causes a freshly
created XRWebGLLayer to return a frame buffer size of 1 x 1 until the first
frame is rendered.

Initializing the left & right viewports with the default framebuffer size fixes
this issue since WebXROpaqueFramebuffer is using the union of the viewports to
report the framebuffer size to the user.

(m_framebufferSize is the size of the setup target for rendering. Starting with
an initial size of 0 x 0 causes the setup logic to run on first frame)

* Source/WebCore/Modules/webxr/WebXROpaqueFramebufferCocoa.cpp:
(WebCore::WebXROpaqueFramebuffer::WebXROpaqueFramebuffer):