Create a kitchen sink app for JSC unit tests https://bugs.webkit.org/show_bug.cgi?id=299416 rdar://161224657
Reviewed by NOBODY (OOPS!).

This introduces a simple framework and a kitchen sink for miscellaneous unit tests. As a
realistic use case, the patch includes a couple of tests for WebAssemblyCompileOptions
which I wished I could write back when I worked on them.

See `SelfTests.cpp` for details and `WebAssemblyTests.cpp` for a couple of real tests.

I am currently working on getting this build properly in all EWS configurations.

SaferCPP complains about code stolen from jsc.cpp, so these are are not new issues.

The style checker complains about unsafety of printf and strncmp used in the test runner.
I claim immunity because the test runner is not a shipping product, and in this case the
usage is safe and the code is simpler this way.