Modulemaps on all platforms https://bugs.webkit.org/show_bug.cgi?id=299424 rdar://161232274
Reviewed by NOBODY (OOPS!).

Modulemaps allow the bmalloc and WTF headers to be treated as clang modules,
which means they can be recorded as abstract syntax trees on disk to avoid the
need to parse them in each compiler invocation. Within the WebKit build, this
is currently only used for Swift compilation.

Previously, we were only putting such modulemaps in place for OS X and iOS
builds. Extend modulemaps across all platforms.