Hard code WTF modulemap https://bugs.webkit.org/show_bug.cgi?id=299427 rdar://problem/161234996
Reviewed by NOBODY (OOPS!).

The clang modulemap for WTF was previously generated by a script; this is no
longer necessary so this PR removes the script and puts in place a hard-coded
modulemap. This is an engineering simplification with (almost) no intentional
functional changes.

Background: when we added the WTF modulemap, the Swift compiler was unable
to successfully parse a variety of WTF headers and so we put each header into
its own mini submodule. This could only be achieved by generating the
modulemap dynamically.

Over time, it became clear this wasn't the right approach. Analysis showed
that the Swift compiler actually only had difficulties with a small number of
WTF headers (possibly the compiler improved) and it started to cause friction
for developers to include the precise header they needed, so in
https://github.com/WebKit/WebKit/pull/50358
we combined the vast majority of the headers into one submodule.

As a happy side-effect, this means we no longer need the generation script
and can instead use
  umbrella "."
to indicate that the modulemap should pick up all the headers in its
directory. That's what this change does. The resulting hard-coded modulemap
is virtually the same as the previous generated modulemap aside from
replacing a long list of headers with a single 'umbrella' statement.