[WPE] Expose PlatformDisplayDefault https://bugs.webkit.org/show_bug.cgi?id=299526
Reviewed by NOBODY (OOPS!).

There are devices where the GBM path for PlatformDisplay
cannot be used due to API limitations, but we can still
use WPE with DMABufs on those with the PlatformDisplayDefault
path. This was already exposed for GTK, now do it for WPE.

A missed bit in AcceleratedSurface was also added to correctly
set the m_type as necessary, either for DMABufs when present
in EGL extensions, or not.

* Source/WebCore/PlatformWPE.cmake:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/platform/graphics/PlatformDisplay.h:
* Source/WebCore/platform/graphics/egl/PlatformDisplayDefault.cpp:
* Source/WebCore/platform/graphics/egl/PlatformDisplayDefault.h:
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/AcceleratedSurface.cpp:
(WebKit::AcceleratedSurface::SwapChain::SwapChain):
* Source/WebKit/WebProcess/glib/WebProcessGLib.cpp:
(WebKit::WebProcess::initializePlatformDisplayIfNeeded const):