[SaferCPP] Lambda capture issues in WebCore Editor/Navigator https://bugs.webkit.org/show_bug.cgi?id=299540 rdar://161348962
Reviewed by NOBODY (OOPS!).

Editor.cpp warning was:
"Implicitly captured raw-pointer 'this' to uncounted type is unsafe".
Fixing by explicitly capturing this pointer and wrapping it in Ref.

Navigator.cpp warning was:
"Captured raw-pointer 'this' to uncounted type is unsafe".
Fixing by capturing a WeakPtr to this and converting it to a RefPtr
with null check when the lambda is invoked.

* Source/WebCore/SaferCPPExpectations/UncountedLambdaCapturesCheckerExpectations:
* Source/WebCore/editing/Editor.cpp:
(WebCore::Editor::scanSelectionForTelephoneNumbers):
* Source/WebCore/page/Navigator.cpp:
(WebCore::Navigator::share):