[GFC] Initial implementation of grid item layout. https://bugs.webkit.org/show_bug.cgi?id=300062 rdar://problem/161860162
Reviewed by NOBODY (OOPS!).

This patch provides the initial implementation of grid item layout as
described in the final step of the "Grid Layout Algorithm,":
https://drafts.csswg.org/css-grid-1/#layout-algorithm

At this point we have resolved the grid area sizes for each grid item.
As a result, we have a definite containing block size for the grid item
so we can resolve the size of the item if needed to provide as
constraints for layout of its contents. For now we keep this as simple
as possible since we will only allow content in which the grid item has
a fixed width and size. This is done via some utility functions that
determine the used width and heights of the items and then pass these
values into our integration API as constraints for the box's formatting
context.