Expose WebKit::NavigationActionData::isRequestFromClientOrUserInput through WKNavigationAction https://bugs.webkit.org/show_bug.cgi?id=300035 <rdar://problem/161833541>
Reviewed by NOBODY (OOPS!).

The NavigationActionData class contains a flag that hints whether an action was triggered
through User Input or a load directive from the User Agent Chrome
(NavigationActionData::isRequestFromClientOrUserInput) (e.g., Bug 266354).

It would be helpful to expose this in the `WKNavigationAction` layer so that the User
Interface could use this as a hint about how verbose to be during certain load failures.
For example, a failed load of a custom URL scheme triggered by script while loading or
updating a page doesn't generally need to be reported to the user, since they aren't aware
of the load attempt and console logging will let a web developer see this. On the other
hand, if a user clicks on a link or enters a URL in the browser chrome and the load fails,
we want to immediately present this information so the user can understand what went wrong.

This patch exposes the existing WebKit::NavigationActionData::isRequestFromClientOrUserInput
flag through the WKNavigationAction class so a client can access it. The data is already
present in the UIProcess as internal state.

Tested by API tests.

* Source/WebKit/UIProcess/API/APINavigationAction.h:
* Source/WebKit/UIProcess/API/Cocoa/WKNavigationAction.mm:
(-[WKNavigationAction _isRequestFromClientOrUserInput]):
* Source/WebKit/UIProcess/API/Cocoa/WKNavigationActionPrivate.h:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/DecidePolicyForNavigationAction.mm:
(TEST(WebKit, DecidePolicyForNavigationActionReload)):
(TEST(WebKit, DecidePolicyForNavigationActionForTargetedHyperlink)):
(runDecidePolicyForNavigationActionForHyperlinkThatRedirects):