[WebGL] texSubImage2D should use video fast path https://bugs.webkit.org/show_bug.cgi?id=300359 rdar://132814893
Reviewed by NOBODY (OOPS!).

When uploading the entire rectangle of video frames from HTMLVideoElement to a
WebGL texture of the same size via texSubImage2D, WebKit performs a slow
conversion of the pixels on CPU, instead of using the fast-path.

The fast path for uploading video frames to WebGL textures was restricted
to only texImage2D calls. This restriction is unnecessary because
sourceImageRectIsDefault already tracks that the entire video frame is
required without change, which applies to both texImage2D and texSubImage2D.

Added PerformanceTests to catch regressions.

* PerformanceTests/Canvas/texImage2DVideo.html: Added.
* PerformanceTests/Canvas/texSubImage2DVideo.html: Added.
* PerformanceTests/Canvas/resources/bip-bop-4300x4300.mp4: Added.
* PerformanceTests/resources/webgl-utils.js: Added.
* Source/WebCore/html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::texImageSource): Remove texImage2D-only restriction.