Add some testing for the "kill ring", including the two different yank commands rdar://162449097 https://bugs.webkit.org/show_bug.cgi?id=300567
Reviewed by NOBODY (OOPS!).

Test: editing/kill-ring.html

* LayoutTests/editing/kill-ring-expected.txt: Added.
* LayoutTests/editing/kill-ring.html: Added.

* Source/WebCore/testing/Internals.cpp:
(WebCore::Internals::executeEditingCommand): Added.
* Source/WebCore/testing/Internals.h: Added executeEditingCommand.
* Source/WebCore/testing/Internals.idl: Ditto.

* Source/WebKit/UIProcess/mac/WebViewImpl.mm:
(WebKit::WebViewImpl::firstRectForCharacterRange): Removed assertion that was hit
while I was using the WebEditingTester. It's not correct to assert something is
not reached based on wishful thinking! There may be work to do here at some point,
but an assertion is the wrong tool.

* Tools/WebEditingTester/WK2WebDocumentController.m:
(defaultConfiguration): Switched from private interface to public.

* Tools/WebEditingTester/WebEditingTester.xcodeproj/project.pbxproj:
Added -Wno-deprecated to the classic WebKit part of this tool to keep it compiling.