[webkitcorepy] subprocess_utils.py (run) should have the ability to stream and capture output. https://bugs.webkit.org/show_bug.cgi?id=300938 rdar://162813867
Reviewed by NOBODY (OOPS!).

A deficiency of `subprocess.run` is that it can't stream output to stdout while
the process is running AND capture that output for use elsewhere. webkitcorepy
has an existing function in subprocess_utils.py (`run`) that wraps
`subprocess.run`. This change adds a new parameter to that function that
streams and captures stdout.

* Tools/Scripts/libraries/webkitcorepy/webkitcorepy/subprocess_utils.py:
    (run): Add parameter `capture_and_stream_output`, change to using `Popen` instead of `run`.