Introduce WTF_OVERRIDE_DELETE_FOR_CHECKED_PTR_NONNULL https://bugs.webkit.org/show_bug.cgi?id=301349 rdar://163263301
Reviewed by NOBODY (OOPS!).

Swift requires C++ raw pointer function arguments and return values to be
annotated with _Nonnull or _Nullable. This is analysed on a file-by-file
granularity. For the upcoming CoreIPC work, we will need to annotate
Source/WebKit/UIProcess/WebProcessProxy.h. This class uses a macro
WTF_OVERRIDE_DELETE_FOR_CHECKED_PTR which expands to create an `operator
delete` which needs to be duly annotated. Unfortunately, we can't just change
WTF_OVERRIDE_DELETE_FOR_CHECKED_PTR always to use _Nonnull, since that would
require us to fully annotate all files in which that macro is expanded.
Instead, we need to create a new version of this macro which expands to NONNULL
only in those header files which we have converted.

Subsequent commits will start to use this new version of the macro in
WebProcessProxy.h.