Event timing: implement targetSelector https://bugs.webkit.org/show_bug.cgi?id=301597 rdar://163598158
Reviewed by NOBODY (OOPS!).

Implements the feature as described by the current spec draft: a string
that is similar to a CSS selector is stored inside the Event Timing
entry generated when finalizeEventTimingEntry() is called.

Adding one extra data member to PerformanceEventTimingCandidate made
it too large for a HashMap, so it was necessary to change how
m_pendingKeyDowns is stored: we now have two separate Vector, one
for keys (m_pendingKeys) and one for values (m_pendingKeyDowns). This
causes lookup to be O(n), but the map is so small that this isn't a
problem - the map contains currently pressed keyboard keys.

No new tests (OOPS!).

* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::computeInteractionID):
(WebCore::LocalDOMWindow::generateCSSSelector):
(WebCore::LocalDOMWindow::initializeEventTimingEntry):
(WebCore::LocalDOMWindow::finalizeEventTimingEntry):
(WebCore::LocalDOMWindow::dispatchPendingEventTimingEntries):
* Source/WebCore/page/LocalDOMWindow.h:
* Source/WebCore/page/PerformanceEventTiming.cpp:
(WebCore::PerformanceEventTiming::PerformanceEventTiming):
* Source/WebCore/page/PerformanceEventTiming.h:
* Source/WebCore/page/PerformanceEventTiming.idl:
* Source/WebCore/page/PerformanceEventTimingCandidate.h: