Reject incompatible CSS Typed OM value types for color properties https://bugs.webkit.org/show_bug.cgi?id=301782 rdar://163122497
Reviewed by NOBODY (OOPS!).

Fix crash that occurs in RenderThemeMac::systemColor() when CSS Typed
OM's attributeStyleMap.set() is used to set a color property with a
BorderImageWidthValue. The crash occurs because the incompatible value
gets misinterpreted as a KeywordColor with valueID=0 (CSSValueInvalid),
and attempting to look up 0 in the systemStyleColors HashMap triggers
an assertion failure since 0 is the empty value sentinel for integer keys.

Fix by adding type validation in StylePropertyMap::set() to reject
BorderImageWidthValue for color properties, throwing a TypeError instead
of crashing.

Test: fast/css/css-typed-om/style-property-map-color-type-validation.html

* LayoutTests/fast/css/css-typed-om/style-property-map-color-type-validation-expected.txt: Added.
* LayoutTests/fast/css/css-typed-om/style-property-map-color-type-validation.html: Added.
* Source/WebCore/css/typedom/StylePropertyMap.cpp:
(WebCore::StylePropertyMap::set):