[ iOS Release ] http/tests/geolocation/geolocation-get-current-position-does-not-leak.https.html is a flakey text failure https://bugs.webkit.org/show_bug.cgi?id=295751 rdar://155570673
Reviewed by NOBODY (OOPS!).

The issue was that the test would call `setInterval()` once per iframe
but would only cancel the setInterval for one frame in case of success.
As a result, setInterval could fire for other iframes after calling
finishJSTest() and log further text.

Address the issue by using a single setInterval for all the frames.

* LayoutTests/http/tests/geolocation/geolocation-get-current-position-does-not-leak.https.html:
* LayoutTests/platform/ios/TestExpectations: