BBQJIT::emitTailCall locked register failure on 32 bit https://bugs.webkit.org/show_bug.cgi?id=302460
Reviewed by NOBODY (OOPS!).

In BBQJIT on 32 bit, when there are no free registers for ScratchScope to select a free register not in a preserved set of registers,
the register allocator may try to free a register that is in a GPR2 pair with a register in the preserved set.
When then BBQJIT::flushValue tries to flush the value contained in the GPR2 pair, it tries to spill both registers in the GPR2 pair,
failing with "ASSERTION FAILED: !m_spiller.isLocked(reg)", because ScratchScope locks the registers in the preserved set.
We see these failures in JSTests/wasm/stress/tail-call.js on ARM_THUMB2 systems.

If we add the missing GPR2 pairs to the preserved set, we end up with failures due to all the registers being locked.

With this patch, store the FP/SP to the stack, as is done in BBQJIT::emitIndirectTailCall for X86_64, and additionally store the LR.
Adjust the sizes of the stack areas for the ARM_THUMB2 case, and append I64 values last.

This fixes JSTests/wasm/stress/tail-call.js on ARM_THUMB2 systems.

This patch was co-authored by Mikhail R. Gadelha <mikhail@igalia.com>.

No new tests (OOPS!).
* Source/JavaScriptCore/wasm/WasmBBQJIT.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::emitTailCall):
(JSC::Wasm::BBQJITImpl::BBQJIT::emitIndirectTailCall):