Safari takes a long time to load a page with hundreds of audio elements (fast in Chrome) https://bugs.webkit.org/show_bug.cgi?id=302371 rdar://158490207
Reviewed by NOBODY (OOPS!).

When we insert a media element into the DOM we will take a few actions which can cause style and layout thrashing
when inserting many media elements. This can lead to multi-minute page loads when there are a few hundred media elements
with UA controls in the document. In contrast, Chrome can handle the load in about a second on a M4 max MacBook Pro.

The actions that are taken are the following:

*) HTMLMediaElement is inserted into the DOM. When it is finished, we configure controls.
This will parse (if necessary) the controls JavaScript code and call createControls.

*) createControls will append to the UA shadow root multiple style elements given the UA stylesheets attached to
the MediaControlsHost. These styles are re-parsed for each HTMLMediaElement inserted.

*) createControls will create the shadow DOM tree for the controls and attach to the shadow root. As part of this
we need to compute their size which ends up walking up the DOM from the media element to the root looking for transforms.
This causes us to thrash style for each media node inserted.

This change does two major things and some minor clean up to account for the new behavior.

First, we cache the UA style sheet in
the document media element's script world. This way media elements which are added to the DOM can adopt the stylesheets instead of
having to re-parse them. As measured by the PerformanceTests test added this takes the total time down from ~180 seconds to ~110 seconds
on my M4 Max MacBook Pro.

Second, when a HTMLMediaElement is added to the DOM we will queue a task to configure the media controls rather than calling the
controls JS code synchronously. This has a massive impact on the amount of style and layout calculations we need to do when adding
media elements. As measured by the PerformanceTests test added here it takes the total time down from ~110 seconds to 4 seconds.

In all, the new PerformanceTests test added measures a 98% reduction in the time it takes to add 1000 audio elements to a page.

Test coverage is via existing tests and a new PerformanceTests test.

* LayoutTests/media/modern-media-controls/:
  Many tests expect controls to be available as soon as the
  media element is inserted into the DOM. Now we must wait.

* LayoutTests/platform/mac/TestExpectations:
  Unskip a passing test.

* PerformanceTests/Media/AudioElementControlsCreation.html: Added.
   Stress test inserting 1000 audio elements with controls.

* Source/WebCore/Modules/modern-media-controls/main.js:
(ensureUAStyleSheets):
(createControls):
  The same UA stylesheets are added as style elements to each media element
  created. This is a lot of redundant CSS parsing that can be cached.

* Source/WebCore/Modules/modern-media-controls/controls/text-tracks.css:
([useragentpart="-webkit-media-text-track-container"]):
(video::-webkit-media-text-track-container): Deleted.
(video[controls]::-webkit-media-text-track-container.visible-controls-bar): Deleted.
  Since the stylesheets are now adopted we need to change the selector to look for useragentpart
  rather than a pseudo.

* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::didFinishInsertingNode):
(WebCore::HTMLMediaElement::layoutSizeChanged):
  Instead of adding controls as soon as the node is inserted into the DOM we will
  queue a task to add the controls, if necessary.

* Source/WebCore/accessibility/cocoa/AXCoreObjectCocoa.mm:
(WebCore::AXCoreObject::isEmptyGroup):
  The accessibility role shouldn't depend on the media element's subtree containing controls or not.

* Source/WebCore/css/html.css:
(audio):
  The UA stylesheet for audio elements had a different default width than the style for an audio element
  with controls. Now that we're asynchronously adding the controls this can cause a web-visible layout shift
  (covered in imported/w3c/web-platform-tests/html/rendering/replaced-elements/embedded-content/audio-controls-intrinsic-size.html).
  So we change the width for audio elements to match the width for audio elements with controls in this stylesheet.